[#9] mine 타입 적용
