+++ src/main/java/io/jaeyeon/cloudboxserver/file/controller/FileApiController.java
@@ -1,10 +1,9 @@
 package io.jaeyeon.cloudboxserver.file.controller;
 
-import io.jaeyeon.cloudboxserver.file.dto.DownloadRequestDto;
 import io.jaeyeon.cloudboxserver.file.dto.DownloadResponseDto;
 import io.jaeyeon.cloudboxserver.file.dto.UploadRequestDto;
 import io.jaeyeon.cloudboxserver.file.dto.UploadResponseDto;
-import io.jaeyeon.cloudboxserver.file.service.S3Service;
+import io.jaeyeon.cloudboxserver.file.service.FileService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -15,22 +14,22 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class FileApiController {
 
-  private final S3Service s3Service;
+  private final FileService fileService;
 
   @GetMapping("/presigned-url")
   public ResponseEntity<UploadResponseDto> getUploadPresignedUrl(
-          @RequestParam("fileName") String fileName,
-          @RequestParam("extension") String extension,
-          @RequestParam("contentType") String contentType) {
+      @RequestParam("fileName") String fileName,
+      @RequestParam("extension") String extension,
+      @RequestParam("contentType") String contentType) {
     UploadRequestDto requestDto = new UploadRequestDto(fileName, extension, contentType);
-    UploadResponseDto responseDto = s3Service.generatePresignedUrl(requestDto);
+    UploadResponseDto responseDto = fileService.generatePresignedUrl(requestDto);
     return ResponseEntity.ok(responseDto);
   }
 
   @GetMapping("/download-url")
   public ResponseEntity<DownloadResponseDto> getDownloadPresignedUrl(
-          @RequestParam("fileName") String fileName) {
-    DownloadResponseDto responseDto = s3Service.generateDownloadPresignedUrl(fileName);
+      @RequestParam("fileName") String fileName) {
+    DownloadResponseDto responseDto = fileService.generateDownloadPresignedUrl(fileName);
     return ResponseEntity.ok(responseDto);
   }
 }
