+++ src/main/java/io/jaeyeon/cloudboxserver/file/controller/FileController.java
@@ -1,11 +1,10 @@
 package io.jaeyeon.cloudboxserver.file.controller;
 
 import io.jaeyeon.cloudboxserver.file.domain.entity.FileEntity;
-import io.jaeyeon.cloudboxserver.file.dto.DownloadRequestDto;
 import io.jaeyeon.cloudboxserver.file.dto.DownloadResponseDto;
 import io.jaeyeon.cloudboxserver.file.dto.UploadRequestDto;
 import io.jaeyeon.cloudboxserver.file.dto.UploadResponseDto;
-import io.jaeyeon.cloudboxserver.file.service.S3Service;
+import io.jaeyeon.cloudboxserver.file.service.FileService;
 import java.io.IOException;
 import java.net.URL;
 import java.util.List;
@@ -21,14 +20,14 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 @RequestMapping("/files")
 public class FileController {
 
-  private final S3Service s3Service;
+  private final FileService fileService;
 
   @GetMapping("/")
   public String listFiles(
-          @RequestParam(defaultValue = "0") int page,
-          @RequestParam(defaultValue = "10") int size,
-          Model model) {
-    List<FileEntity> files = s3Service.listFiles(page, size);
+      @RequestParam(defaultValue = "0") int page,
+      @RequestParam(defaultValue = "10") int size,
+      Model model) {
+    List<FileEntity> files = fileService.listFiles(page, size);
     model.addAttribute("files", files);
     model.addAttribute("currentPage", page);
     return "fileList";
@@ -36,27 +35,25 @@ public class FileController {
 
   @PostMapping("/upload")
   public String uploadFile(
-          @RequestParam("file") MultipartFile file, RedirectAttributes redirectAttributes)
-          throws IOException {
+      @RequestParam("file") MultipartFile file, RedirectAttributes redirectAttributes)
+      throws IOException {
 
     String fileName = file.getOriginalFilename();
 
     String extension = fileName.contains(".") ? fileName.substring(fileName.lastIndexOf('.')) : "";
-    UploadRequestDto requestDto = new UploadRequestDto(
-            fileName, extension, file.getContentType());
-    UploadResponseDto responseDto = s3Service.generatePresignedUrl(requestDto);
+    UploadRequestDto requestDto = new UploadRequestDto(fileName, extension, file.getContentType());
+    UploadResponseDto responseDto = fileService.generatePresignedUrl(requestDto);
 
-    s3Service.uploadToS3(requestDto, file, new URL(responseDto.presignedUrl()));
+    fileService.uploadToS3(requestDto, file, new URL(responseDto.presignedUrl()));
 
     redirectAttributes.addFlashAttribute(
-            "message", "File uploaded successfully: " + file.getOriginalFilename());
+        "message", "File uploaded successfully: " + file.getOriginalFilename());
     return "redirect:/files/";
   }
 
   @GetMapping("/download")
-  public String downloadFile(
-          @RequestParam("fileName") String fileName) {
-    DownloadResponseDto responseDto = s3Service.generateDownloadPresignedUrl(fileName);
+  public String downloadFile(@RequestParam("fileName") String fileName) {
+    DownloadResponseDto responseDto = fileService.generateDownloadPresignedUrl(fileName);
     return "redirect:" + responseDto.presignedUrl();
   }
 }
