+++ src/main/java/io/jaeyeon/cloudboxserver/file/domain/entity/FileEntity.java
@@ -1,11 +1,6 @@
 package io.jaeyeon.cloudboxserver.file.domain.entity;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -32,13 +27,14 @@ public class FileEntity {
   private String path;
 
   @Column(nullable = false)
-  private String fileType;
+  @Enumerated(EnumType.STRING)
+  private FileType mine;
 
   @Builder
-  public FileEntity(String fileName, long size, String path, String fileType) {
+  public FileEntity(String fileName, Long size, String path, FileType mine) {
     this.fileName = fileName;
     this.size = size;
     this.path = path;
-    this.fileType = fileType;
+    this.mine = mine;
   }
 }
