+++ src/main/java/io/jaeyeon/cloudboxserver/file/dto/DownloadRequestDto.java
@@ -3,5 +3,4 @@ package io.jaeyeon.cloudboxserver.file.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
-public record DownloadRequestDto(
-        @NotBlank @Size(max = 255) String fileName) { }
+public record DownloadRequestDto(@NotBlank @Size(max = 255) String fileName) {}
