+++ src/main/java/io/jaeyeon/cloudboxserver/file/dto/DownloadResponseDto.java
@@ -1,3 +1,3 @@
 package io.jaeyeon.cloudboxserver.file.dto;
 
-public record DownloadResponseDto(String presignedUrl) { }
+public record DownloadResponseDto(String presignedUrl) {}
