+++ src/main/java/io/jaeyeon/cloudboxserver/file/dto/UploadRequestDto.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
 public record UploadRequestDto(
-        @NotBlank @Size(max = 255) String fileName,
-        @NotBlank @Size(max = 10) String extension,
-        @NotBlank @Size(max = 50) String contentType) {}
+    @NotBlank @Size(max = 255) String fileName,
+    @NotBlank @Size(max = 10) String extension,
+    @NotBlank @Size(max = 50) String contentType) {}
